Move callbacks to AuthManager
